Item column and calendar logic update for UX refactor

+++ actions/items/deleteItem.js
@@ -1,7 +1,6 @@
 'use server';
 
 import Task from '../../models/Task';
-//import { revalidatePath } from 'next/cache';
 import { handleServerErrorMessage } from '../../utilities';
 import { getUserFromCookie } from '../../utilities/getUserFromCookie';
 
@@ -18,9 +17,13 @@ export default async function deleteTask(_id, userId) {
   }
 
   try {
-    const result = await Task.findOne({ _id: _id });
-    await Task.deleteOne({ _id: _id });
-    // if (isCalendarItem) revalidatePath('/dashboard');
+    const result = await Task.findOne({ _id: _id, userId: cookieUserId });
+
+    if (!result) {
+      return { status: 404, error: 'Item not found' };
+    }
+
+    await Task.deleteOne({ _id: _id, userId: cookieUserId });
     return { status: 200, item: JSON.parse(JSON.stringify(result)) };
   } catch (error) {
     const errorMessage = handleServerErrorMessage(error);
